task7 projet15 v11

+++ pointers_arrays_strings/8-print_diagsums_save_save_6sur7.c
@@ -8,7 +8,7 @@
 void print_diagsums(int *a, int size)
 {
 int i;
-long sum1 =0, sum2 = 0;
+int sum1 =0, sum2 = 0;
 
 	for (i = 0; i < size; i++)
 	{
@@ -16,7 +16,7 @@ long sum1 =0, sum2 = 0;
 	sum2 += a[i * size + (size - 1 - i)]; /* Sum of the secondary diagonal */
 	}
 
-	printf("%ld, %ld\n", sum1, sum2);
+	printf("%d, %d\n", sum1, sum2);
 
 	/*fprintf(sum1);
 	_putchar (',');
@@ -24,3 +24,4 @@ long sum1 =0, sum2 = 0;
 	fprintf(sum2);
 	_putchar ('\n');*/
 }
+
